refactor(config): Extract interfaces from "index.ts" to "types.ts"
