+++ lib/config/massage.spec.ts
@@ -1,6 +1,6 @@
 import { getName } from '../../test/util';
 import * as massage from './massage';
-import { RenovateConfig } from '.';
+import type { RenovateConfig } from './types';
 
 describe(getName(__filename), () => {
   describe('massageConfig', () => {
