+++ lib/util/package-rules.spec.ts
@@ -1,4 +1,4 @@
-import type { PackageRuleInputConfig, UpdateType } from '../config';
+import type { PackageRuleInputConfig, UpdateType } from '../config/types';
 import {
   LANGUAGE_DOCKER,
   LANGUAGE_JAVASCRIPT,
