+++ lib/workers/branch/automerge.spec.ts
@@ -1,6 +1,6 @@
 import { defaultConfig, getName, git, platform } from '../../../test/util';
-import { RenovateConfig } from '../../config';
 import { setAdminConfig } from '../../config/admin';
+import type { RenovateConfig } from '../../config/types';
 import { BranchStatus } from '../../types';
 import { tryBranchAutomerge } from './automerge';
