+++ lib/workers/repository/init/vulnerability.ts
@@ -1,4 +1,4 @@
-import { PackageRule, RenovateConfig } from '../../../config';
+import type { PackageRule, RenovateConfig } from '../../../config/types';
 import { NO_VULNERABILITY_ALERTS } from '../../../constants/error-messages';
 import * as datasourceMaven from '../../../datasource/maven';
 import { id as npmId } from '../../../datasource/npm';
