+++ lib/workers/repository/onboarding/pr/base-branch.ts
@@ -1,4 +1,4 @@
-import { RenovateConfig } from '../../../../config';
+import type { RenovateConfig } from '../../../../config/types';
 
 export function getBaseBranchDesc(config: RenovateConfig): string {
   // Describe base branch only if it's configured
